Seems like the commit message was deemed empty and my last commit didn't work. So:
Added: Pressing the add button makes the Edit Object Dialog box open
Added: Made sure that alt tabbing to program puts the correct window on top.
Added: Window is now properly sized, but DPI is not yet taken into account.
Added: Window is now centered on the monitor it is on.
Added: Ok and Cancel buttons, and their strings on the dialog.