Removed more dead code.
Fixed the issue with editing descriptions using the wrong method.

Made it so the base dialog no longer needs to enable it's parent window
when closing. This is to handle the edge case where a dialog fails to
be created and thus is unable to restore it's parents previous state.