Deleted some more dead code.
Made it so that if a dialog does indeed fail to create itself, the
dialog is cleaned up first before displaying an error.

Had to revert the previous change regarding enabling the parent window.
Doing it the other way resulted in the window and dialog flickering.

The Window is now notified when the tile is updated, how it was.